% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eruptR-package.R
\docType{package}
\name{eruptR-package}
\alias{eruptR}
\alias{eruptR-package}
\title{eruptR: Exploring Earth’s Fiery Giants}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Akanksha Shridharani \email{ashr0017@student.monash.edu}

}
\keyword{internal}
