% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explosive_index.R
\name{explosive_index}
\alias{explosive_index}
\title{Volcano Name by their highest Volcanic Explosivity Index (VEI)}
\usage{
explosive_index(start_year)
}
\arguments{
\item{start_year}{A numeric input greater than 46}
}
\description{
A function which takes the year of the volcano as an input and returns the volcano name and
its highest Volcanic Explosivity Index for that specific year.
}
\examples{
explosive_index(2011)
}
