% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci_sequence.R
\name{fibonacci_series}
\alias{fibonacci_series}
\title{Generation of a Fibonacci Sequence}
\usage{
fibonacci_series(num)
}
\arguments{
\item{num}{A numeric input greater than 0}
}
\description{
A function which takes an integer value as an input and
it will generate a Fibonacci sequence as an output. It is a sequence to generate a sequence
in which each of the number is a value of the sum of the two previous numbers.
}
\examples{
fibonacci_series(10)
}
