% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eruption_counts_data.R
\docType{data}
\name{new_data}
\alias{new_data}
\title{Dataset about volcanic eruptions by their Explosivity Index for recent 20 yeats}
\format{
A data frame with 127 rows and 3 variables.
\describe{
\item{volcano_name}{Volcano Name}
\item{vei}{Volcano explosivity Index}
\item{start_year}{Volcano Start Year}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-05-12/eruptions.csv}
}
\usage{
new_data
}
\description{
A dataset that contains the the different types of volcanic eruption according to their eruption, which is used in
a function related to R-shiny to display the Volcanic Explosivity Index and its respective volcano according to year.
}
\keyword{datasets}
