% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{run_app}
\alias{run_app}
\title{R-Shiny Application on different Volcanoes}
\usage{
run_app()
}
\description{
This functions runs the R-shiny application created in Assignment 1
for the eruptR package.
}
