% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_counts_data.R
\docType{data}
\name{volcano_height}
\alias{volcano_height}
\title{Dataset about different types of volcanoes}
\format{
A data frame with 70 rows and 26 variables.
\describe{
\item{volcano_number}{Volcano unique ID}
\item{volcano_name}{Volcano Name}
\item{primary_volcano_type	}{Types of Volcano}
\item{last_eruption_year}{Last Year Erupted}
\item{country}{Country}
\item{region}{Region}
\item{subregion}{Sub Region}
\item{latitude}{Latitude}
\item{longitude}{Longitude}
\item{elevation}{Elevation}
\item{tectonic_settings}{Volcano Plate Tectonic Settings}
\item{evidence_category}{Type of Evidence}
\item{major_rock_1}{Major Rock Type}
\item{major_rock_2}{Major Rock Type}
\item{major_rock_3}{Major Rock Type}
\item{major_rock_4}{Major Rock Type}
\item{major_rock_5}{Major Rock Type}
\item{minor_rock_1}{Minor Rock Type}
\item{minor_rock_2}{Minor Rock Type}
\item{minor_rock_3}{Minor Rock Type}
\item{minor_rock_4}{Minor Rock Type}
\item{minor_rock_5}{Minor Rock Type}
\item{population_within_5_km}{Total population within 5 km of volcano}
\item{population_within_10_km}{Total population within 10 km of volcano}
\item{population_within_30_km}{Total population within 30 km of volcano}
\item{population_within_100_km}{Total population within 100 km of volcano}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-05-12/volcano.csv}
}
\usage{
volcano_height
}
\description{
A dataset that contains the the different types of volcanoes worldwide, their elevation showing only the first 70 observations in descending order
to help us knowing whether its a nature's fury or geological wonder
}
\keyword{datasets}
