% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanic_eruption_counts.R
\name{volcanic_eruption_counts}
\alias{volcanic_eruption_counts}
\title{Generation of Volcanic Eruptions for each year by Volcanic Explosivity Index}
\usage{
volcanic_eruption_counts(year_input)
}
\arguments{
\item{year_input}{The year parameter which is used for plotting}
}
\description{
A parameterised function which takes an input of R-shiny one panel function i.e year
as an input and it will generate us Volcanic Explosivity Index(VEI) vs Volcano Name.
It mainly tell us that more VEI means that that the eruption of that volcano is really powerful.
The data has been selected to show the top 7 highest VEI volcanoes in the last 20 years.
}
\examples{
volcanic_eruption_counts(2001)
}
